+++ adapter/persistence/src/main/java/org/anastaklo/portal/adapter/persistence/PersonRepository.java
@@ -1,0 +1,4 @@
+package org.anastaklo.portal.adapter.persistence;
+
+public interface PersonRepository {
+}
