+++ adapter/web/pom.xml
@@ -4,27 +4,19 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <packaging>jar</packaging>
-    <artifactId>application</artifactId>
-    
-    <parent>
-        <groupId>org.anastaklo</groupId>
-        <artifactId>employee-portal</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
     <dependencies>
         <dependency>
             <groupId>org.anastaklo</groupId>
             <artifactId>api-in</artifactId>
             <version>${portal.module.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.modelmapper</groupId>
-            <artifactId>modelmapper</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter</artifactId>
-        </dependency>
     </dependencies>
+
+    <artifactId>portal-web</artifactId>
+
+    <parent>
+        <groupId>org.anastaklo</groupId>
+        <artifactId>employee-portal</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
 </project>
