+++ adapter/web/src/main/java/org/anastaklo/portal/adapter/web/GetPersonalDataController.java
@@ -1,0 +1,8 @@
+package org.anastaklo.portal.adapter.web;
+
+import org.anastaklo.portal.port.in.GetPersonalDataQuery;
+
+public class GetPersonalDataController {
+    
+    GetPersonalDataQuery getQuery;
+}
