+++ api/api-in/src/main/java/org/anastaklo/portal/port/in/GetPersonalDataQuery.java
@@ -1,0 +1,6 @@
+package org.anastaklo.portal.port.in;
+
+public interface GetPersonalDataQuery {
+
+    PersonDTO getPersonalData(long personId);
+}
