+++ api/api-in/src/main/java/org/anastaklo/portal/port/in/PersonDTO.java
@@ -1,0 +1,4 @@
+package org.anastaklo.portal.port.in;
+
+public class PersonDTO {
+}
