+++ application/src/main/java/org/anastaklo/portal/application/mapper/Mapper.java
@@ -1,0 +1,10 @@
+package org.anastaklo.portal.application.mapper;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+
+class Mapper {
+    
+    @Autowired
+    ModelMapper mapper;
+}
