+++ application/src/main/java/org/anastaklo/portal/application/service/GetPersonalDataService.java
@@ -1,0 +1,12 @@
+package org.anastaklo.portal.application.service;
+
+import org.anastaklo.portal.port.in.GetPersonalDataQuery;
+import org.anastaklo.portal.port.in.PersonDTO;
+
+class GetPersonalDataService implements GetPersonalDataQuery {
+    @Override
+    public PersonDTO getPersonalData(long personId) {
+
+        return null;
+    }
+}
