--- application/src/main/java/org/anastaklo/portal/service/GetPersonalDataService.java
@@ -1,4 +1,0 @@
-package org.anastaklo.portal.service;
-
-public class GetPersonalDataService {
-}
