+++ configuration/src/main/java/org/anastaklo/portal/configuration/PortalApplication.java
@@ -1,0 +1,13 @@
+package org.anastaklo.portal.configuration;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PortalApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PortalApplication.class, args);
+	}
+
+}
