+++ configuration/src/main/java/org/anastaklo/portal/configuration/PortalConfiguration.java
@@ -1,0 +1,14 @@
+package org.anastaklo.portal.configuration;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class PortalConfiguration {
+    
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+}
