+++ domain/src/main/java/org/anastaklo/portal/entities/Department.java
@@ -1,0 +1,5 @@
+package org.anastaklo.portal.entities;
+
+public enum Department {
+    IT, FINANCE, HR
+}
