+++ domain/src/main/java/org/anastaklo/portal/entities/Person.java
@@ -1,0 +1,18 @@
+package org.anastaklo.portal.entities;
+
+import java.time.LocalDate;
+import lombok.Data;
+
+@Data
+public class Person {
+    
+    private String firstName;
+    private String lastName;
+    private LocalDate birthDate;
+    private String email;
+    private String phone;
+    private LocalDate startWorkDate;
+    private LocalDate finishWorkDate;
+    private Department department;
+    private boolean isActive;
+}
