+++ pom.xml
@@ -4,31 +4,38 @@
     <modelVersion>4.0.0</modelVersion>
     <packaging>pom</packaging>
     <modules>
-        <module>adapters/portal-persistence</module>
-        <module>adapters/portal-web</module>
-        <module>api-in</module>
-        <module>api-out</module>
+        <module>adapter/persistence</module>
+        <module>adapter/web</module>
+        <module>api/api-in</module>
+        <module>api/api-out</module>
         <module>application</module>
         <module>domain</module>
-        <module>portal-configuration</module>
+        <module>configuration</module>
     </modules>
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
         <version>2.2.4.RELEASE</version>
-        <relativePath/>
     </parent>
-    <groupId>org.anastaklo</groupId>
-    <artifactId>employee-portal</artifactId>
-    <version>1.0-SNAPSHOT</version>
-
     <properties>
         <java.version>13</java.version>
-        <start-class>org.anastaklo.portal.PortalApplication</start-class>
+        <portal.module.version>1.0-SNAPSHOT</portal.module.version>
     </properties>
+    
+    <groupId>org.anastaklo</groupId>
+    <artifactId>employee-portal</artifactId>
+    <version>${portal.module.version}</version>
 
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>org.modelmapper</groupId>
+                <artifactId>modelmapper</artifactId>
+                <version>2.3.6</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
     <dependencies>
-
         <dependency>
             <groupId>org.postgresql</groupId>
             <artifactId>postgresql</artifactId>
