--- portal-application/src/main/java/org/anastaklo/portal/domain/Department.java
@@ -1,5 +1,0 @@
-package org.anastaklo.portal.domain;
-
-public enum Department {
-    IT, FINANCE, HR
-}
