--- portal-configuration/src/main/java/org/anastaklo/portal/PortalApplication.java
@@ -1,13 +1,0 @@
-package org.anastaklo.portal;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PortalApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PortalApplication.class, args);
-	}
-
-}
